Add sass-lint, refactor components into index.js
