+++ daemon/web/.sass-lint.yml
@@ -1,0 +1,11 @@
+files:
+  include:
+    - '**/*.sass'
+  ignore:
+    - 'node_modules/**/*'
+    - 'public/**/*'
+
+rules:
+  single-line-per-selector: 0
+  nesting-depth: 0
+  no-transition-all: 0
