+++ daemon/web/components/Footer/index.js
