+++ daemon/web/components/Footer/index.sass
@@ -1,44 +1,47 @@
-@import ../../styles/index
-   
+@import '../../styles/index'
+
 .footer
-    display: flex
-    flex-flow: row
-    width: 100%
-    height: 4rem
-    align-items: center
-    justify-content: space-between
+  align-items: center
+  display: flex
+  flex-flow: row
+  height: 4rem
+  justify-content: space-between
+  width: 100%
 
-    .icons 
-        font-size: 20px
-        object-fit: contain
-        padding: 0.5rem
+  .icons
+    font-size: 20px
+    object-fit: contain
+    padding: .5rem
 
-        *
-            display: inline-block
+    *
+      display: inline-block
 
-        > div
-            height: 100%
-            vertical-align: middle
-        > a
-            color: $lightblue
-            &:hover
-                color: $lightred
+      > div
+        height: 100%
+        vertical-align: middle
+
+      > a
+        color: $lightblue
+
+        &:hover
+          color: $lightred
 
 .launchpad
-     width: 105px
-     height: 100%
-     padding-left: 0.1rem
-     fill: $lightblue
-     &:hover
-        fill: $lightred
+  fill: $lightblue
+  height: 100%
+  padding-left: .1rem
+  width: 105px
+
+  &:hover
+    fill: $lightred
 
 .right-side
-    padding-right: 10px
+  padding-right: 10px
 
 .left-side
-    padding-left: 18px
+  padding-left: 18px
 
 .daemon-version
-    color: $lightblue
-    letter-spacing: 1px
-    font-size: 10px
+  color: $lightblue
+  font-size: 10px
+  letter-spacing: 1px
