+++ daemon/web/components/IconHeader/index.sass
@@ -1,13 +1,13 @@
 @import '../../styles/index'
 
-.header
-    color: $grey
-    display: inline
-    margin-left: 10px;
-    font-weight: bold;
-
 i
-    color: $grey
+  color: $grey
+
+.header
+  color: $grey
+  display: inline
+  font-weight: bold
+  margin-left: 10px
 
 .iconheader
-    margin: 30px 0 30px 40px
+  margin: 30px 0 30px 40px
