+++ daemon/web/components/LogView/index.js
