--- daemon/web/components/PruneImageButton/index.sass
@@ -1,14 +1,0 @@
-@import '../../styles/index'
-
-.PruneImageButton
-  font-family: Rubik
-  font-size: 12px
-  font-weight: 500
-  letter-spacing: 1px
-  padding: 5px 10px
-  border-radius: 4px
-  border: none
-  border-bottom: 1px solid $separator
-  
-  &:disabled
-    opacity: 0.5
