+++ daemon/web/components/Status/index.js
