+++ daemon/web/components/Status/index.sass
@@ -1,21 +1,20 @@
 @import '../../styles/index'
 
 .badge
-    color: $grey
-    display: inline-block
+  color: $grey
+  display: inline-block
 
 .title
-    display: inline-block
-    margin: 0 10px 0 40px;
+  display: inline-block
+  margin: 0 10px 0 40px
 
-.badge-pill-active
-    padding: 2px 10px
-    border-radius: 100px
-    background-image: linear-gradient(257deg, #BAFFD7, #BAEFE8)
-    font-size: 14px
+.badge-pill
+  border-radius: 100px
+  font-size: 14px
+  padding: 2px 10px
 
-.badge-pill-inactive
-    padding: 2px 10px
-    border-radius: 100px
-    background-image: linear-gradient(257deg, #E4E4E4, #CACACA)
-    font-size: 14px
+  &-active
+    background-image: linear-gradient(257deg, $gradient-active-start, $radient-active-end)
+
+  &-inactive
+    background-image: linear-gradient(257deg, $gradient-inactive-start, $gradient-inactive-end)
