+++ daemon/web/components/Table/index.js
