+++ daemon/web/components/buttons/ShutdownButton/index.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import './index.sass';
+import '../index.sass';
 
 const ShutdownButton = ({ style }) => (
   <div>
-    <button className="ShutdownButton" type="button" style={style}>
+    <button className="button" type="button" style={style}>
 Shut Down Containers
     </button>
   </div>
