+++ daemon/web/components/buttons/index.js
@@ -1,0 +1,7 @@
+import PruneImageButton from './PruneImageButton';
+import ShutdownButton from './ShutdownButton';
+
+export {
+  PruneImageButton,
+  ShutdownButton,
+};
