+++ daemon/web/components/buttons/index.sass
@@ -1,14 +1,14 @@
 @import '../../styles/index'
 
-.ShutdownButton
+.button
+  border: 0
+  border-bottom: 1px solid $separator
+  border-radius: 4px
   font-family: Rubik
   font-size: 12px
   font-weight: 500
   letter-spacing: 1px
   padding: 5px 10px
-  border-radius: 4px
-  border: none
-  border-bottom: 1px solid $separator
-  
+
   &:disabled
-    opacity: 0.5
+    opacity: .5
