+++ daemon/web/index.html
@@ -9,7 +9,7 @@
 </head>
 
 <body>
-  <div id="app"></div>
+  <div id="app" class="app"></div>
 </body>
 
 </html>
