+++ daemon/web/pages/containers/Containers.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import TerminalView from '../../components/TerminalView/TerminalView';
-import IconHeader from '../../components/IconHeader/IconHeader';
-import Status from '../../components/Status/Status';
+import TerminalView from '../../components/TerminalView';
+import IconHeader from '../../components/IconHeader';
+import Status from '../../components/Status';
 import './index.sass';
 
 const mocklogs = [
@@ -27,7 +27,7 @@ class ContainersWrapper extends React.Component {
     return (
       <div>
         <IconHeader type="containers" title="/inertia-deploy-test_dev_1" />
-        <div className="containerInfo">
+        <div className="container-info">
           <Status title="Status:" status="Active" />
           <h3>
 Last Updated:
