+++ daemon/web/pages/containers/index.sass
@@ -1,14 +1,15 @@
 @import '../../styles/index'
 
-.containerInfo h3
-  display: inline-block
-  margin-left: 40px
+.container-info
+  > h3
+    display: inline-block
+    margin-left: 40px
 
-.containerInfo h4
-  display: inline-block
-  margin-left: 10px
-  margin-bottom: 20px
+  > h4
+    display: inline-block
+    margin-bottom: 20px
+    margin-left: 10px
 
-.terminalView
-  width: 1375px
+.terminal-view
   height: 496px
+  width: 1375px
