+++ daemon/web/styles/colors.sass
@@ -1,15 +1,28 @@
-/* Color palette */
+// Color palette
 
+$white: #fff
+$offwhite: #212b36
 $greyblue: #666c7b
 $darkgrey: #212b36
 $grey: #474d5e
 $lightblue: #d3fafb
-$lightred: #F76C6C
+$lightred: #f76c6c
 $selection-highlight: #e5f0fc
 $nav-select: #bbecf9
+
+$background: #f9f9f9
+$a-color: #5ac4ba
+$a-hover-color: #00a09f
+$separator: rgba(151, 151, 151, .14)
+
+$shadow: rgba(155, 155, 155, .14)
+
+// Gradients
 $gradient-lightblue: #dcffeb
 $gradient-green: #bcf8ab
-$background: #f9f9f9
-$a-color: #5AC4BA
-$a-hover-color: #00A09F
-$separator: rgba(151, 151, 151, 0.14)
+
+$gradient-active-start: #baffd7
+$gradient-active-end: #baefe8
+
+$gradient-inactive-start: #e4e4e4
+$gradient-inactive-end: #cacaca
