+++ daemon/web/styles/typography.sass
@@ -1,34 +1,31 @@
-/* Text styles */
+// Text styles
 
 h1
+  color: $grey
   font-family: Rubik
   font-size: 24px
   font-weight: 500
   letter-spacing: 1px
-  color: $grey
 
 h2
+  color: $greyblue
   font-family: Rubik
   font-size: 16px
   font-weight: 500
   letter-spacing: 1px
-  color: $greyblue
 
 h3
+  color: $darkgrey
   font-family: Rubik
   font-size: 14px
   font-weight: 500
-  color: $darkgrey
 
 h4
+  color: $greyblue
   font-family: Rubik
   font-size: 14px
-  color: $greyblue
 
 p
+  color: $white
   font-family: Rubik
   font-size: 10px
-  color: #ffffff
-
-
-
